feat: add SyncLocalStorage
새로고침시 localstorage를 동기화 하는 로직을 분리 및 적용
