+++ src/pages/_app.tsx
@@ -1,12 +1,16 @@
 import '@/styles/globals.css';
-import type { AppProps } from 'next/app';
 
 import { RecoilRoot } from 'recoil';
+import type { AppProps } from 'next/app';
+
+import SyncLocalStorage from '@/components/SyncLocalStorage';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <RecoilRoot>
-      <Component {...pageProps} />
+      <SyncLocalStorage>
+        <Component {...pageProps} />
+      </SyncLocalStorage>
     </RecoilRoot>
   );
 }
